Changes to files in placement folder to improve portability and eliminate type-conversion warnings and problems compiling code that makes use of Open MP. size_t becomes intptr_t in many places, so that omp directives will work with compilers that are fussier about the index variables in #pragma omp parallel for loops being signed (e.g. in ParallelParsimonyCalculator).

+++ placement/parallelparsimonycalculator.h
@@ -67,11 +67,11 @@ public:
      @param start_index the "top" of the stack (the point at which to begin calculation)
      @param taskDescription a description of the task; may point to an empty string
      */
-    void calculate(int start_index = 0, const char* taskDescription="");
+    void calculate(intptr_t start_index = 0, const char* taskDescription="");
     
-    int parsimonyLink4Cost(PhyloNode* a, PhyloNode* b, PhyloNode* c,
-                           PhyloNode* d, PhyloNode* e, PhyloNode* f,
-                           UINT* buffer1, UINT* buffer2);
+    double parsimonyLink4Cost(PhyloNode* a, PhyloNode* b, PhyloNode* c,
+                              PhyloNode* d, PhyloNode* e, PhyloNode* f,
+                              UINT* buffer1, UINT* buffer2);
 
 };
